pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "peptak"
version = "0.1.0"
description = "Peptide permeability modeling with MD-GAK / PMD-GAK kernels and Gaussian Processes"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [{name="Your Name", email="you@example.com"}]
dependencies = [
  "numpy>=1.23",
  "pandas>=2.0",
  "scikit-learn>=1.2",
  "tqdm>=4.66",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "joblib>=1.3", 
  # only try rdkit-pypi on Python < 3.12
  "rdkit-pypi>=2022.9.5; python_version < '3.12'"
]

[project.optional-dependencies]
# if someone on py<=3.11 wants to explicitly ask for rdkit via pip
rdkit = ["rdkit-pypi>=2022.9.5; python_version < '3.12'"]

[project.scripts]
peptak = "peptak.cli:main"

[tool.setuptools.packages.find]
where = ["src"]


